+++ openpyxl/writer/styles.py
@@ -1,3 +1,4 @@
+from __future__ import absolute_import
 # Copyright (c) 2010-2014 openpyxl
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
@@ -24,34 +25,29 @@
 """Write the shared style table."""
 
 # package imports
-from openpyxl.shared.xmltools import Element, SubElement
-from openpyxl.shared.xmltools import get_document_content
-from openpyxl.shared.ooxml import SHEET_MAIN_NS
-from openpyxl import style
+from openpyxl.xml.functions import (
+    Element,
+    SubElement,
+    ConditionalElement,
+    get_document_content,
+    )
+from openpyxl.compat import safe_string
+from openpyxl.xml.constants import SHEET_MAIN_NS
+
+from openpyxl.styles import DEFAULTS, Protection
+from openpyxl.styles.fills import GradientFill, PatternFill
+
 
 class StyleWriter(object):
 
     def __init__(self, workbook):
-        self._style_list = self._get_style_list(workbook)
+        self.workbook = workbook
         self._style_properties = workbook.style_properties
-        self._root = Element('styleSheet',
-            {'xmlns':'http://schemas.openxmlformats.org/spreadsheetml/2006/main'})
+        self._root = Element('styleSheet', {'xmlns': SHEET_MAIN_NS})
 
-    def _get_style_list(self, workbook):
-        crc = {}
-        for worksheet in workbook.worksheets:
-            uniqueStyles = dict((id(style), style) for style in worksheet._styles.values()).values()
-            for style in uniqueStyles:
-                crc[hash(style)] = style
-        self.style_table = dict([(style, i + 1) \
-            for i, style in enumerate(crc.values())])
-        sorted_styles = sorted(self.style_table.items(), \
-            key=lambda pair:pair[1])
-        return [s[0] for s in sorted_styles]
-
-    def get_style_by_hash(self):
-        return dict([(hash(style), id) \
-            for style, id in self.style_table.items()])
+    @property
+    def styles(self):
+        return self.workbook.shared_styles
 
     def write_table(self):
         number_format_table = self._write_number_formats()
@@ -65,6 +61,24 @@ class StyleWriter(object):
         self._write_table_styles()
 
         return get_document_content(xml_node=self._root)
+
+    def _unpack_color(self, node, color, key='color'):
+        """Convert colors encoded as RGB, theme or tints
+        Possible values
+        RGB: #4F81BD
+        Theme: theme:9
+        Tint: theme:9:7 # guess work
+        """
+        if color is None:
+            return
+        if not ":" in color:
+            SubElement(node, key, {'rgb': color})
+        else:
+            _, theme, tint = color.split(":")
+            if tint == '':
+                SubElement(node, key, {'theme': theme})
+            else:
+                SubElement(node, key, {'theme': theme, 'tint': tint})
 
     def _write_fonts(self):
         """ add fonts part to root
@@ -84,30 +98,20 @@ class StyleWriter(object):
         # others
         table = {}
         index = 1
-        for st in self._style_list:
-            if hash(st.font) != hash(style.DEFAULTS.font) and hash(st.font) not in table:
-                table[hash(st.font)] = str(index)
+        for st in self.styles:
+            if st.font != DEFAULTS.font and st.font not in table:
+                table[st.font] = index
                 font_node = SubElement(fonts, 'font')
                 SubElement(font_node, 'sz', {'val':str(st.font.size)})
-                if str(st.font.color.index).split(':')[0] == 'theme': # strip prefix theme if marked as such
-                    if str(st.font.color.index).split(':')[2]:
-                        SubElement(font_node, 'color', {'theme':str(st.font.color.index).split(':')[1],
-                                                        'tint':str(st.font.color.index).split(':')[2]})
-                    else:
-                        SubElement(font_node, 'color', {'theme':str(st.font.color.index).split(':')[1]})
-                else:
-                    SubElement(font_node, 'color', {'rgb':str(st.font.color.index)})
+                self._unpack_color(font_node, st.font.color.index)
                 SubElement(font_node, 'name', {'val':st.font.name})
                 SubElement(font_node, 'family', {'val':'2'})
                 # Don't write the 'scheme' element because it appears to prevent
                 # the font name from being applied in Excel.
                 #SubElement(font_node, 'scheme', {'val':'minor'})
-                if st.font.bold:
-                    SubElement(font_node, 'b')
-                if st.font.italic:
-                    SubElement(font_node, 'i')
-                if st.font.underline == 'single':
-                    SubElement(font_node, 'u')
+                ConditionalElement(font_node, "b", st.font.bold)
+                ConditionalElement(font_node, "i", st.font.italic)
+                ConditionalElement(font_node, "u",  st.font.underline == 'single')
 
                 index += 1
 
@@ -123,34 +127,34 @@ class StyleWriter(object):
 
         table = {}
         index = 2
-        for st in self._style_list:
-            if hash(st.fill) != hash(style.DEFAULTS.fill) and hash(st.fill) not in table:
-                table[hash(st.fill)] = str(index)
-                fill = SubElement(fills, 'fill')
-                if hash(st.fill.fill_type) != hash(style.DEFAULTS.fill.fill_type):
-                    node = SubElement(fill, 'patternFill', {'patternType':st.fill.fill_type})
-                    if hash(st.fill.start_color) != hash(style.DEFAULTS.fill.start_color):
-                        if str(st.fill.start_color.index).split(':')[0] == 'theme': # strip prefix theme if marked as such
-                            if str(st.fill.start_color.index).split(':')[2]:
-                                SubElement(node, 'fgColor', {'theme':str(st.fill.start_color.index).split(':')[1],
-                                                             'tint':str(st.fill.start_color.index).split(':')[2]})
-                            else:
-                                SubElement(node, 'fgColor', {'theme':str(st.fill.start_color.index).split(':')[1]})
-                        else:
-                            SubElement(node, 'fgColor', {'rgb':str(st.fill.start_color.index)})
-                    if hash(st.fill.end_color) != hash(style.DEFAULTS.fill.end_color):
-                        if str(st.fill.end_color.index).split(':')[0] == 'theme': # strip prefix theme if marked as such
-                            if str(st.fill.end_color.index).split(':')[2]:
-                                SubElement(node, 'bgColor', {'theme':str(st.fill.end_color.index).split(':')[1],
-                                                             'tint':str(st.fill.end_color.index).split(':')[2]})
-                            else:
-                                SubElement(node, 'bgColor', {'theme':str(st.fill.end_color.index).split(':')[1]})
-                        else:
-                            SubElement(node, 'bgColor', {'rgb':str(st.fill.end_color.index)})
+        for st in self.styles:
+            if st.fill != DEFAULTS.fill and st.fill not in table:
+
+                table[st.fill] = index
+                node = SubElement(fills, 'fill')
+                if isinstance(st.fill, PatternFill):
+                    self._write_pattern_fill(node, st.fill)
+                elif isinstance(st.fill, GradientFill):
+                    self._write_gradient_fill(node, st.fill)
                 index += 1
 
         fills.attrib["count"] = str(index)
         return table
+
+    def _write_pattern_fill(self, node, fill):
+        if fill != DEFAULTS.fill and fill.fill_type is not None:
+            node = SubElement(node, 'patternFill', {'patternType':
+                                                    fill.fill_type})
+            if fill.start_color != DEFAULTS.fill.start_color:
+                self._unpack_color(node, fill.start_color.index, 'fgColor')
+            if fill.end_color != DEFAULTS.fill.end_color:
+                self._unpack_color(node, fill.end_color.index, 'bgColor')
+
+    def _write_gradient_fill(self, node, fill):
+        node = SubElement(node, 'gradientFill', dict(fill))
+        for idx, color in enumerate(fill.stop):
+            stop = SubElement(node, "stop", {"position":safe_string(idx)})
+            self._unpack_color(stop, color.index)
 
     def _write_borders(self):
         borders = SubElement(self._root, 'borders')
@@ -166,26 +170,19 @@ class StyleWriter(object):
         # others
         table = {}
         index = 1
-        for st in self._style_list:
-            if hash(st.borders) != hash(style.DEFAULTS.borders) and hash(st.borders) not in table:
-                table[hash(st.borders)] = str(index)
+        for st in self.styles:
+            if st.borders != DEFAULTS.borders and st.borders not in table:
+                table[st.borders] = index
                 border = SubElement(borders, 'border')
                 # caution: respect this order
                 for side in ('left', 'right', 'top', 'bottom', 'diagonal'):
                     obj = getattr(st.borders, side)
                     if obj.border_style is None or obj.border_style == 'none':
                         node = SubElement(border, side)
-                        attrs = {}
                     else:
                         node = SubElement(border, side, {'style':obj.border_style})
-                        if str(obj.color.index).split(':')[0] == 'theme': # strip prefix theme if marked as such
-                            if str(obj.color.index).split(':')[2]:
-                                SubElement(node, 'color', {'theme':str(obj.color.index).split(':')[1],
-                                                                'tint':str(obj.color.index).split(':')[2]})
-                            else:
-                                SubElement(node, 'color', {'theme':str(obj.color.index).split(':')[1]})
-                        else:
-                            SubElement(node, 'color', {'rgb':str(obj.color.index)})
+                        self._unpack_color(node, obj.color.index)
+
                 index += 1
 
         borders.attrib["count"] = str(index)
@@ -201,50 +198,51 @@ class StyleWriter(object):
 
         # writing the cellXfs
         cell_xfs = SubElement(self._root, 'cellXfs',
-            {'count':'%d' % (len(self._style_list) + 1)})
+            {'count':'%d' % (len(self.styles) + 1)})
 
         # default
         def _get_default_vals():
             return dict(numFmtId='0', fontId='0', fillId='0',
-                xfId='0', borderId='0')
+                        xfId='0', borderId='0')
 
-        SubElement(cell_xfs, 'xf', _get_default_vals())
-
-        for st in self._style_list:
+        for st in self.styles:
             vals = _get_default_vals()
 
-            if hash(st.font) != hash(style.DEFAULTS.font):
-                vals['fontId'] = fonts_table[hash(st.font)]
+            if st.font != DEFAULTS.font:
+                vals['fontId'] = str(fonts_table[st.font])
                 vals['applyFont'] = '1'
 
-            if hash(st.borders) != hash(style.DEFAULTS.borders):
-                vals['borderId'] = borders_table[hash(st.borders)]
+            if st.borders != DEFAULTS.borders:
+                vals['borderId'] = str(borders_table[st.borders])
                 vals['applyBorder'] = '1'
 
-            if hash(st.fill) != hash(style.DEFAULTS.fill):
-                vals['fillId'] = fills_table[hash(st.fill)]
+            if st.fill != DEFAULTS.fill:
+                vals['fillId'] = str(fills_table[st.fill])
                 vals['applyFill'] = '1'
 
-            if st.number_format != style.DEFAULTS.number_format:
+            if st.number_format != DEFAULTS.number_format:
                 vals['numFmtId'] = '%d' % number_format_table[st.number_format]
                 vals['applyNumberFormat'] = '1'
 
-            if hash(st.alignment) != hash(style.DEFAULTS.alignment):
+            if st.alignment != DEFAULTS.alignment:
                 vals['applyAlignment'] = '1'
+
+            if st.protection != DEFAULTS.protection:
+                vals['applyProtection'] = '1'
 
             node = SubElement(cell_xfs, 'xf', vals)
 
-            if hash(st.alignment) != hash(style.DEFAULTS.alignment):
+            if st.alignment != DEFAULTS.alignment:
                 alignments = {}
 
                 for align_attr in ['horizontal', 'vertical']:
-                    if hash(getattr(st.alignment, align_attr)) != hash(getattr(style.DEFAULTS.alignment, align_attr)):
+                    if getattr(st.alignment, align_attr) != getattr(DEFAULTS.alignment, align_attr):
                         alignments[align_attr] = getattr(st.alignment, align_attr)
 
-                    if hash(st.alignment.wrap_text) != hash(style.DEFAULTS.alignment.wrap_text):
+                    if st.alignment.wrap_text != DEFAULTS.alignment.wrap_text:
                         alignments['wrapText'] = '1'
 
-                    if hash(st.alignment.shrink_to_fit) != hash(style.DEFAULTS.alignment.shrink_to_fit):
+                    if st.alignment.shrink_to_fit != DEFAULTS.alignment.shrink_to_fit:
                         alignments['shrinkToFit'] = '1'
 
                     if st.alignment.indent > 0:
@@ -257,6 +255,20 @@ class StyleWriter(object):
 
                 SubElement(node, 'alignment', alignments)
 
+            if st.protection != DEFAULTS.protection:
+                protections = {}
+
+                if st.protection.locked == Protection.PROTECTION_PROTECTED:
+                    protections['locked'] = '1'
+                elif st.protection.locked == Protection.PROTECTION_UNPROTECTED:
+                    protections['locked'] = '0'
+
+                if st.protection.hidden == Protection.PROTECTION_PROTECTED:
+                    protections['hidden'] = '1'
+                elif st.protection.hidden == Protection.PROTECTION_UNPROTECTED:
+                    protections['hidden'] = '0'
+
+                SubElement(node, 'protection', protections)
 
     def _write_cell_style(self):
         cell_styles = SubElement(self._root, 'cellStyles', {'count':'1'})
@@ -268,70 +280,41 @@ class StyleWriter(object):
             dxfs = SubElement(self._root, 'dxfs', {'count': str(len(self._style_properties['dxf_list']))})
             for d in self._style_properties['dxf_list']:
                 dxf = SubElement(dxfs, 'dxf')
-                if 'font' in d and d['font']:
+                if 'font' in d and d['font'] is not None:
                     font_node = SubElement(dxf, 'font')
-                    if 'color' in d['font']:
-                        if str(d['font']['color'].index).split(':')[0] == 'theme':  # strip prefix theme if marked
-                            if str(d['font']['color'].index).split(':')[2]:
-                                SubElement(font_node, 'color', {'theme': str(d['font']['color'].index).split(':')[1],
-                                                                'tint': str(d['font']['color'].index).split(':')[2]})
-                            else:
-                                SubElement(font_node, 'color', {'theme': str(d['font']['color'].index).split(':')[1]})
-                        else:
-                            SubElement(font_node, 'color', {'rgb': str(d['font']['color'].index)})
-                    # Don't write the 'scheme' element because it appears to prevent
-                    # the font name from being applied in Excel.
-                    #SubElement(font_node, 'scheme', {'val':'minor'})
-                    if 'bold' in d['font'] and d['font']['bold']:
-                        SubElement(font_node, 'b')
-                    if 'italic' in d['font'] and d['font']['italic']:
-                        SubElement(font_node, 'i')
-                    if 'underline' in d['font'] and d['font']['underline'] == 'single':
-                        SubElement(font_node, 'u')
-                if 'fill' in d and len(d['fill']):
-                    f = d['fill'][0]
+                    if d['font'].color is not None:
+                        self._unpack_color(font_node, d['font'].color.index)
+                    ConditionalElement(font_node, 'b', d['font'].bold, 'val')
+                    ConditionalElement(font_node, 'i', d['font'].italic, 'val')
+                    ConditionalElement(font_node, 'u', d['font'].underline != 'none',
+                                       {'val': d['font'].underline})
+                    ConditionalElement(font_node, 'strike', d['font'].strikethrough)
+
+
+                if 'fill' in d:
+                    f = d['fill']
                     fill = SubElement(dxf, 'fill')
                     if f.fill_type:
                         node = SubElement(fill, 'patternFill', {'patternType': f.fill_type})
                     else:
                         node = SubElement(fill, 'patternFill')
-                    if hash(f.start_color) != hash(style.DEFAULTS.fill.start_color):
-                        if str(f.start_color.index).split(':')[0] == 'theme':  # strip prefix theme if marked
-                            if str(f.start_color.index).split(':')[2]:
-                                SubElement(node, 'fgColor', {'theme': str(f.start_color.index).split(':')[1],
-                                                             'tint': str(f.start_color.index).split(':')[2]})
-                            else:
-                                SubElement(node, 'fgColor', {'theme': str(f.start_color.index).split(':')[1]})
-                        else:
-                            SubElement(node, 'fgColor', {'rgb': str(f.start_color.index)})
-                    if hash(f.end_color) != hash(style.DEFAULTS.fill.end_color):
-                        if str(f.end_color.index).split(':')[0] == 'theme':  # strip prefix theme if marked
-                            if str(f.end_color.index).split(':')[2]:
-                                SubElement(node, 'bgColor', {'theme': str(f.end_color.index).split(':')[1],
-                                                             'tint': str(f.end_color.index).split(':')[2]})
-                            else:
-                                SubElement(node, 'bgColor', {'theme': str(f.end_color.index).split(':')[1]})
-                        else:
-                            SubElement(node, 'bgColor', {'rgb': str(f.end_color.index)})
-                if 'border' in d and len(d['border']):
-                    borders = d['border'][0]
+                    if f.start_color != DEFAULTS.fill.start_color:
+                        self._unpack_color(node, f.start_color.index, 'fgColor')
+
+                    if f.end_color != DEFAULTS.fill.end_color:
+                        self._unpack_color(node, f.end_color.index, 'bgColor')
+
+                if 'border' in d:
+                    borders = d['border']
                     border = SubElement(dxf, 'border')
                     # caution: respect this order
-                    for side in ('left', 'right', 'top', 'bottom', 'diagonal'):
+                    for side in ('left', 'right', 'top', 'bottom'):
                         obj = getattr(borders, side)
                         if obj.border_style is None or obj.border_style == 'none':
                             node = SubElement(border, side)
-                            attrs = {}
                         else:
                             node = SubElement(border, side, {'style': obj.border_style})
-                            if str(obj.color.index).split(':')[0] == 'theme':  # strip prefix theme if marked as such
-                                if str(obj.color.index).split(':')[2]:
-                                    SubElement(node, 'color', {'theme': str(obj.color.index).split(':')[1],
-                                                               'tint': str(obj.color.index).split(':')[2]})
-                                else:
-                                    SubElement(node, 'color', {'theme': str(obj.color.index).split(':')[1]})
-                            else:
-                                SubElement(node, 'color', {'rgb': str(obj.color.index)})
+                            self._unpack_color(node, obj.color.index)
         else:
             dxfs = SubElement(self._root, 'dxfs', {'count': '0'})
         return dxfs
@@ -351,7 +334,7 @@ class StyleWriter(object):
         num_fmt_id = 165 # start at a greatly higher value as any builtin can go
         num_fmt_offset = 0
 
-        for style in self._style_list:
+        for style in self.styles:
 
             if not style.number_format in number_format_list  :
                 number_format_list.append(style.number_format)